Vulkan: Move definition of page to `templates`
